+++ packages/core/types/src/configuration.ts
@@ -208,10 +208,6 @@ export interface JWTVerifyOptions extends VerifyOptions {}
 
 export interface APITokenOptions {
   legacy: boolean;
-  /**
-   * Temporary flag to allow migration to the new legacy signature
-   */
-  migrateToSecureLegacySignature: boolean;
   jwt?: JWTOptions;
 }