+++ packages/proxy/test/proxy.metadata.spec.ts
@@ -9,7 +9,7 @@ import type { Logger } from '@verdaccio/types';
 
 import { ProxyStorage } from '../src';
 
-const getConf = (name) => path.join(__dirname, '/conf', name);
+const getConf = (name) => path.join(import.meta.dirname, '/conf', name);
 
 const mockDebug = vi.fn();
 const mockInfo = vi.fn();
@@ -235,7 +235,7 @@ describe('proxy', () => {
         expect(mockInfo).toHaveBeenCalledTimes(2);
         expect(mockInfo).toHaveBeenLastCalledWith(
           {
-            error: 'Response code 500 (Internal Server Error)',
+            error: `Request failed with status code 500 (Internal Server Error): GET ${domain}/jquery`,
             request: { method: 'GET', url: `${domain}/jquery` },
             retryCount: 2,
           },