+++ packages/web/test/api.search.test.ts
@@ -18,7 +18,7 @@ vi.mock('@verdaccio/ui-theme', () => mockManifest());
 describe('test web server', () => {
   beforeAll(() => {
     mockManifest.mockReturnValue(() => ({
-      staticPath: path.join(__dirname, 'static'),
+      staticPath: path.join(import.meta.dirname, 'static'),
       manifestFiles: {
         js: ['runtime.js', 'vendors.js', 'main.js'],
       },